+++ backend/src/main/java/com/allog/dallog/domain/composition/application/RegisterService.java
@@ -1,4 +1,4 @@
-package com.allog.dallog.application;
+package com.allog.dallog.domain.composition.application;
 
 import com.allog.dallog.domain.auth.dto.OAuthMember;
 import com.allog.dallog.domain.category.application.CategoryService;
