+++ backend/src/test/java/com/allog/dallog/domain/composition/application/CategorySubscriptionServiceTest.java
@@ -1,4 +1,4 @@
-package com.allog.dallog.application;
+package com.allog.dallog.domain.composition.application;
 
 import static com.allog.dallog.common.fixtures.CategoryFixtures.공통_일정_생성_요청;
 import static com.allog.dallog.common.fixtures.CategoryFixtures.공통_일정_이름;
